refactor: move "client.ts" logic into "microsoft-api/index.ts"
